+Added location saving

+++ app/src/main/java/nz/ac/unitec/easybuy/MapFragment.java
@@ -19,7 +19,6 @@ import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapView;
 import com.google.android.gms.maps.MapsInitializer;
 import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
@@ -108,8 +107,8 @@ public class MapFragment extends Fragment implements OnMapReadyCallback, Product
             String id = product.getId();
             String name = product.getName();
             float price = product.getPrice();
-            float lat = product.getLat();
-            float lng = product.getLng();
+            double lat = product.getLat();
+            double lng = product.getLng();
 
             product.setMarker(mGoogleMap.addMarker(new MarkerOptions().position(new LatLng(lat, lng)).title(name).snippet(lat + ", " + lng)));
             CameraPosition camPos = CameraPosition.builder().target(new LatLng(lat, lng)).zoom(16).bearing(0).tilt(45).build();
